Services/controller/requirements.txt
Django==2.2.4
Pillow
boto3
google-cloud-storage
python-resize-image
dependency-injector